Fix crash occurring when a capture group is optional and isn't matched
When an optional capture group is specified (e.g., "(a)?"), and as a
result doesn't match, the location of the range at that index is
`NSNotFound`. Attempting to build a `Range<String.UTF16Index>` would
then result in an out of bounds error and crash (because `NSNotFound`
is a very large number).

This is a breaking change on the type of `MatchResult.captures`
(`[String]` => `[String?]`). This change could theoretically be made
non-breaking by using the type `[String!]` to allow for convenient
access for the arguably more common case of non-optional capture
groups. This would force the user to explicitly use optional binding in
order to restore safety. However this is not advisable, as this kind of
crash is likely to be surprising. For that reason this change opts for safety.

Usually the code that accesses capture results will be close to the
code that constructs the regular expression, e.g.:

    let regex = Regex("Hello, (\\w+)")

    if let match = regex.match("Hello, world") {
      let subject = match.captures[0]!
      print("Greeted \(subject)")
    }

Here we can clearly see that there is a single capture group, and it is
non-optional. So just as we can assume that it is safe to access
`captures[0]`, we can also safely use the force unwrapping operator to
extract the captured string.