Overload Regex.init to accept `StaticString`
Forcing all usages of `Regex` to be failable, even if the pattern string
is statically known to be a valid regular expression, unfortunate.

The _only_ way to construct a `StaticString` is with a string literal
(or creating an empty one with `StaticString()`). So using
`StaticString` is thus a way to enforce at compile time that the string
has come from a trusted source.

This change reverts `Regex.init(_:)` to be non-failable by making it's
argument a `StaticString`. If a static string isn't a valid regex, it's
programmer error and not bad input, so it's fine to raise a precondition
failure.

The failable form is now `Regex.init(string:)` (modelled after
`NSURL(string:)`), and `throws` rather than returning `nil`. This means
that `Regex.error` can be removed in favour of standard Swift error handling.