Replace usage of String.UTF16View with built in range conversions
Fixes #67.

At some point in the Swift 4 timeline, presumably with SE-0180[1] which
overhauled String's various views and index types, there have been
built-in and rather convenient conversions between `NSRange` and
`Range<String.Index>`.

This change removes all explicit use of `String.UTF16View` and
`String.UTF16Index` in favour of the `Range<String.Index>.init(_:in:)`
initialiser, which accepts an `NSRange` and maps its encoded offsets to
String indices.

This also removes the need for the internal `Memo` type, as being able
to convert directly to String indices removes a level of indirection
that `Memo` was working around.

[1]: https://github.com/apple/swift-evolution/blob/master/proposals/0180-string-index-overhaul.md