Nest Options within the Regex type
Fixes #73.

Because Regex vends a public type with the same name as its module,
there's no way to disambiguate between these two meanings of
`Regex.Options`: it's either a type name `Options` in the `Regex` module
or a type named options in the `Regex.Regex` type. Swift _always_ picks
the latter, so `Regex.Options` does not compile. Additionally, if
another module exports a public `Options` type, the problem gets even
tricker.

To resolve this we officially nest `Options` inside the `Regex` type, so
the notation which previously failed to compile now works. We also
introduce a deprecated top level typealias for the old symbol, directing
users to use canonical name `Regex.Options`.